# api/openapi-spec/swagger.yaml
openapi: 3.1.0
info:
  title: 考核项目API说明文档
  description: 欢迎使用
  contact:
    email: 1598665020@qq.com
  version: 1.0.11
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/admin/users/status:
    patch:
      tags:
        - Admin
      summary: 修改用户状态
      operationId: changeUserStatus
      requestBody:
        content:
          application/json:
            example: {"userId":1000000,"operate":"enable"}
        description: operate => enable 启用, disable 禁用
      responses:
        200:
          description: OK
        400:
          description: Invalid params
        401:
          description: Unauthorized
        403:
          description: 你没有该操作权限
        404:
          description: 用户不存在
  /api/admin/users/passwd:
    patch:
      tags:
        - Admin
      summary: 重置用户密码
      operationId: resetUserPass
      requestBody:
        content:
          application/json:
            example: {"userId":1000000}
      responses:
        200:
          description: OK
        400:
          description: Invalid params
        401:
          description: Unauthorized
        403:
          description: 你没有该操作权限
        404:
          description: 用户不存在
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: admin(admin::adminId::token) user(userId::token)
...